telegram-bot/requirements.txt
requests==2.26.0
python-telegram-bot==13.7
emoji==1.5.0